Fixed users db

+++ myrate/myrate_backend/server/db/conn.js
@@ -24,37 +24,46 @@ module.exports = {
       }
       
       const collection = client.db("media").collection("users");
-      // The drop() command destroys all data from a collection.
-        // Make sure you run it against proper database and collection.
-        try {
-          collection.drop();
+      
+      if(collection.countDocuments() == 0)
+      {
 
-        }
-        catch (err)
-        {
-
-        }
-
-        // make a bunch of time series data
-        let userData = [];
-
-        for (let i = 0; i < 5000; i++) {
-            const firstName = faker.name.firstName();
-            const lastName = faker.name.lastName();
-            let newUser = {
-                firstName,
-                lastName,
-                day_joined: faker.date.past(),
-                email: faker.internet.email(firstName, lastName),
-                username: faker.name.firstName() + faker.random.word() + faker.random.alphaNumeric(),
-                password: faker.internet.password(),
-            };
-
-            userData.push(newUser);
-        }
-        collection.insertMany(userData);
-
-        console.log("User Database seeded");
+        // The drop() command destroys all data from a collection.
+          // Make sure you run it against proper database and collection.
+          try {
+            collection.drop();
+  
+          }
+          catch (err)
+          {
+  
+          }
+  
+          // make a bunch of time series data
+          let userData = [];
+  
+          for (let i = 0; i < 5000; i++) {
+              const firstName = faker.name.firstName();
+              const lastName = faker.name.lastName();
+              let newUser = {
+                  firstName,
+                  lastName,
+                  day_joined: faker.date.past(),
+                  email: faker.internet.email(firstName, lastName),
+                  username: faker.name.firstName() + faker.random.word() + faker.random.alphaNumeric(),
+                  password: faker.internet.password(),
+              };
+  
+              userData.push(newUser);
+          }
+          collection.insertMany(userData);
+  
+          console.log("User are already in database seeded");
+      }
+      else
+      {
+        console.log("Users already in database. Did not seed.")
+      }
 
 
       return callback(err);
